#191: move peer-list screen state from RootComponent into PeerListComponent
pendingFiles, pendingSources, dropFeedback, setPendingFiles, clearPendingFiles,
onPeerTapped, onDropRejectedDuringTransfer, observeAutoSend, setAutoSend all
move from RootComponent to PeerListComponent — they are peer-list-screen state,
not navigation state.

RootComponent now keeps only: the child stack, createChild, showTransferDetails,
and a peerListComponent getter so platform entry points and RootContent can
reach PeerListComponent without going through the stack.

Judgement call: exposing peerListComponent as a getter on RootComponent
(rather than introducing a PendingFilesSink interface) keeps the diff minimal
and avoids premature abstraction — there is exactly one PeerListComponent per
RootComponent lifetime, and platform entry points don't yet call these methods.
