+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/presentation/RootComponent.kt
@@ -6,28 +6,17 @@ import com.arkivanov.decompose.router.stack.StackNavigation
 import com.arkivanov.decompose.router.stack.childStack
 import com.arkivanov.decompose.router.stack.pop
 import com.arkivanov.decompose.router.stack.pushNew
-import com.arkivanov.decompose.value.MutableValue
 import com.arkivanov.decompose.value.Value
 import com.arkivanov.essenty.lifecycle.coroutines.coroutineScope
-import com.tubetoast.tether.preferences.PeerPreferencesStore
 import com.tubetoast.tether.presentation.transfer.TransferDetailsComponent
-import com.tubetoast.tether.transfer.FileSource
 import com.tubetoast.tether.transfer.PeerIdentity
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.launch
 
 class RootComponent(
     componentContext: ComponentContext,
     private val peerListFactory: (ComponentContext, onShowDetails: (PeerIdentity) -> Unit) -> PeerListComponent,
     coroutineScope: CoroutineScope = componentContext.coroutineScope(),
-    private val peerPreferencesStore: PeerPreferencesStore? = null,
 ) : ComponentContext by componentContext {
-    private val scope = coroutineScope
-
     private sealed interface Config {
         data object DeviceList : Config
 
@@ -38,7 +27,10 @@ class RootComponent(
 
     private val navigation = StackNavigation<Config>()
 
-    private var peerListComponent: PeerListComponent? = null
+    private var _peerListComponent: PeerListComponent? = null
+
+    val peerListComponent: PeerListComponent
+        get() = requireNotNull(_peerListComponent) { "PeerListComponent not yet created" }
 
     val stack: Value<ChildStack<*, Child>> = childStack(
         source = navigation,
@@ -48,25 +40,15 @@ class RootComponent(
         childFactory = ::createChild,
     )
 
-    private val mutablePendingFiles = MutableValue(PendingFilesSummary.NONE)
-    val pendingFiles: Value<PendingFilesSummary> = mutablePendingFiles
-
-    private val mutablePendingSources = MutableValue<List<FileSource>>(emptyList())
-
-    private val mutableDropFeedback = MutableValue(false)
-    val dropFeedback: Value<Boolean> = mutableDropFeedback
-
-    private var dropFeedbackJob: Job? = null
-
     private fun createChild(config: Config, context: ComponentContext): Child =
         when (config) {
             Config.DeviceList -> {
                 val plc = peerListFactory(context, ::showTransferDetails)
-                peerListComponent = plc
+                _peerListComponent = plc
                 Child.DeviceListChild(plc)
             }
             is Config.TransferDetails -> {
-                val peerComponent = requireNotNull(peerListComponent?.peerTransferComponent(config.peer)) {
+                val peerComponent = requireNotNull(_peerListComponent?.peerTransferComponent(config.peer)) {
                     "No PeerTransferComponent for ${config.peer} — navigate to device list first"
                 }
                 Child.TransferDetailsChild(
@@ -79,42 +61,8 @@ class RootComponent(
             }
         }
 
-    fun setPendingFiles(summary: PendingFilesSummary, sources: List<FileSource>) {
-        mutablePendingFiles.value = summary
-        mutablePendingSources.value = sources
-    }
-
-    fun clearPendingFiles() {
-        mutablePendingFiles.value = PendingFilesSummary.NONE
-        mutablePendingSources.value = emptyList()
-    }
-
-    fun onPeerTapped(peer: PeerIdentity) {
-        val sources = mutablePendingSources.value
-        if (sources.isEmpty()) return
-        peerListComponent?.peerTransferComponent(peer)?.startOutbound(sources)
-        clearPendingFiles()
-    }
-
     fun showTransferDetails(peer: PeerIdentity) {
         navigation.pushNew(Config.TransferDetails(peer))
-    }
-
-    fun observeAutoSend(peer: PeerIdentity): Flow<Boolean> =
-        peerPreferencesStore?.observeAutoSend(peer) ?: flowOf(false)
-
-    fun setAutoSend(peer: PeerIdentity, enabled: Boolean) {
-        val store = peerPreferencesStore ?: return
-        scope.launch { store.setAutoSend(peer, enabled) }
-    }
-
-    fun onDropRejectedDuringTransfer() {
-        dropFeedbackJob?.cancel()
-        dropFeedbackJob = scope.launch {
-            mutableDropFeedback.value = true
-            delay(DROP_FEEDBACK_DURATION_MS)
-            mutableDropFeedback.value = false
-        }
     }
 
     sealed interface Child {
@@ -125,9 +73,5 @@ class RootComponent(
         data class TransferDetailsChild(
             val component: TransferDetailsComponent,
         ) : Child
-    }
-
-    private companion object {
-        const val DROP_FEEDBACK_DURATION_MS = 3_000L
     }
 }
